refactor(analytics): restructure system overview services and introduce per-section caching
- Split system_overview_helpers into modular service layer:
- Implemented per-section caching
Notes
- Reduced redundant DB queries for overlapping requests)
