+++ docker/nginx/default.conf
@@ -12,13 +12,13 @@ server {
 
     # Normal HTTP traffic
     location / {
-        proxy_pass http://web:8000;
+        proxy_pass http://api:8000;
         proxy_set_header Host $host;
         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
         proxy_set_header X-Forwarded-Proto https;
     }
     location /ws/ {
-        proxy_pass http://web:8000;
+        proxy_pass http://api:8000;
         proxy_http_version 1.1;
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection $connection_upgrade;
